# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ITA Profiles page API documentation",
        "description": "Documentation needed for the API Rest for the ITA Profiles page to IT Academy. Some useful links below:\r\n *   [ITA Profles Backend](https://github.com/IT-Academy-BCN/ita-profiles-backend)\r\n *   [ITA Profiles Frontend](https://github.com/IT-Academy-BCN/ita-profiles-frontend)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1"
        }
    ],
    "paths": {
...
        "/students/{id}": {
            "get": {
                "tags": [
                    "Student"
                ],
                "summary": "Get details of a Student.",
                "description": "Get the details of a specific student. No authentication required",
                "operationId": "getStudentDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the student",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success. Returns student details.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "John"
                                                    },
                                                    "surname": {
                                                        "type": "string",
                                                        "example": "Doe"
                                                    },
                                                    "subtitle": {
                                                        "type": "string",
                                                        "example": "Engineer and Developer."
                                                    },
                                                    "about": {
                                                        "type": "string",
                                                        "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit."
                                                    },
                                                    "cv": {
                                                        "type": "string",
                                                        "example": "My currículum."
                                                    },
                                                    "bootcamp": {
                                                        "type": "string",
                                                        "example": "PHP Developer"
                                                    },
                                                    "endDate": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "..."
                                                    },
                                                    "linkedin": {
                                                        "type": "string",
                                                        "example": "http://www.linkedin.com"
                                                    },
                                                    "github": {
                                                        "type": "string",
                                                        "example": "http://www.github.com"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found."
                    }
                }
            },
            "put": {
                "tags": [
                    "Student"
                ],
                "summary": "Update a Student.",
                "description": "Update the details of a specific User. Requires student or admin role and valid token.",
                "operationId": "updateStudent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Student to be updated.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "subtitle": {
                                        "type": "string",
                                        "example": "Engineer and Full Stack Developer"
                                    },
                                    "bootcamp": {
                                        "type": "enum",
                                        "example": "PHP Developer"
                                    },
                                    "about": {
                                        "type": "text",
                                        "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit."
                                    },
                                    "cv": {
                                        "type": "string",
                                        "example": "Updated Curriculum."
                                    },
                                    "linkedin": {
                                        "type": "string",
                                        "example": "http://www.linkedin.com"
                                    },
                                    "github": {
                                        "type": "string",
                                        "example": "http://www.github.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success. Returns Student details.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "John"
                                        },
                                        "surname": {
                                            "type": "string",
                                            "example": "Doe"
                                        },
                                        "subtitle": {
                                            "type": "string",
                                            "example": "Engineer and Full Stack Developer"
                                        },
                                        "bootcamp": {
                                            "type": "enum",
                                            "example": "PHP Developer"
                                        },
                                        "endDate": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "..."
                                        },
                                        "about": {
                                            "type": "text",
                                            "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit."
                                        },
                                        "cv": {
                                            "type": "string",
                                            "example": "Updated Curriculum."
                                        },
                                        "linkedin": {
                                            "type": "string",
                                            "example": "http://www.linkedin.com"
                                        },
                                        "github": {
                                            "type": "string",
                                            "example": "http://www.github.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Missing authentication token, admin role, student role, or not matching id.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "It was not possible to complete transaction.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong. Try it again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Student"
                ],
                "summary": "Delete a Student.",
                "description": "Delete a specific Student-User by his ID. Requires student or admin role and valid token.",
                "operationId": "deleteStudent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Student to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Student deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Student deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Missing authentication token, admin role, student role, or not matching id.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "It was not possible to complete transaction.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong. Try it again later-"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    }
}